move theme location and references
