+++ src/features/parser/controlsContainer.js
@@ -5,38 +5,13 @@ import Octicon from 'react-octicon'
 import styled from 'styled-components'
 
 import Button from  '../common/button'
-import ParseStatusButton from  '.parseStatusButton'
-
-import { media } from  '../common/mediaQuery'
-import { colour, space } from '../common/theme'
+import Controls from  '../common/controls'
+import ParseStatusButton from  './parseStatusButton'
 
 import {
   addWarrior
 } from './actions'
 
-const Controls = styled.div`
-
-  grid-row-start: 2;
-  border: none;
-
-  ${media.phone`
-    border-top: 1px solid ${colour.grey};
-    background-color: ${colour.lightbg};
-    grid-row-start: 4;
-  `}
-  ${media.tablet`
-    border-top: 1px solid ${colour.grey};
-    background-color: ${colour.lightbg};
-  `}
-  ${media.desktop`
-    border-top: 1px solid ${colour.grey};
-    background-color: ${colour.lightbg};
-  `}
-
-  display: flex;
-  flex-direction: row;
-  height: calc(${space.controls} - 1px);
-`
 
 const MobileControls = ({ addWarrior, currentParseResult }) => (
   <Controls>
