+++ src/features/topbar/tabLink.js
@@ -2,7 +2,7 @@ import styled from 'styled-components'
 import { NavLink } from 'react-router-dom'
 import styledProperty from 'styled-property'
 
-import { colour, space } from '../../styles/theme'
+import { colour, space } from '../common/theme'
 
 const BaseLink = styled(NavLink)`
   color: ${colour.grey};
